+++ src/module/mod.rs
@@ -554,6 +554,7 @@ impl Module {
     #[cfg(not(feature = "no_function"))]
     #[cfg(not(feature = "no_module"))]
     #[inline]
+    #[allow(dead_code)]
     pub(crate) fn set_raw_fn_as_scripted(
         &mut self,
         name: impl Into<String>,